requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0

SQLAlchemy==2.0.32

numpy==2.0.2
opencv-python-headless==4.10.0.84
Pillow==10.4.0

torch>=2.2.0 --index-url https://download.pytorch.org/whl/cpu
facenet-pytorch==2.5.3

bcrypt==3.2.2
python-jose[cryptography]==3.5.0
python-dotenv==1.2.1